Apply to all objects
Should now be compatible with any new object categories added by other mods - e.g. to avoid the IBCPALLETS problem happening again.

N.B. Cannot apply to BALES because otherwise the string for 'Amount' would be the same.  Prefer to use this instead of a custom string because it is already translated into all supported languages